\docType{data}
\name{dataExample}
\alias{dataExample}
\title{Spatial distribution of the Dicorynia guianensis species of trees in the French Guiana terra firme rainforest.}
\format{A list with 3 objects\describe{
	\item{data}{A \code{\link{data.frame}} with two columns \code{x} and \code{y} corresponding the tree locations in the plot.}
	\item{w0}{An object of class \code{\link{owin}} characterizing the window of study.}
	\item{Z.Pente}{An object of class \code{\link{im}} containing the value of a covariate.}
	}
}
\source{
A. Bar-Hen, M. Emily and N. Picard. (2015) Spatial Cluster Detection Using Nearest Neighbour Distance, Spatial Statistics, Vol. 14, pages 400-411.\cr
}
\usage{
data(dataExample)
}
\description{
	The \code{dataExample} object contains tree locations, window of study and the value of a covariate in the plot.
	}
\value{
A dataset containing the tree locations, window of study and the value of a covariate in the plot.}

\keyword{datasets}

