\docType{data}
\name{GenomicRegion}
\alias{GenomicRegion}
\title{Haplotype counts around gene HSA2 in domestic dogs}
\format{A data.frame with 30 rows and 124 columns. Each row is a dog breed and each column is an haplotype.}
\source{
Emily, M., Hitte, C. and Mom, A. SMILE: a novel Dissimilarity-based Procedure for Detecting Sparse-Specific Profiles in Sparse Contingency Tables. Computational Statistics and Data Analysis,  Vol. 99, pages 171-188, 2016.\cr
}
\usage{
data(GenomicRegion)
}
\description{
	A matrix-like dataset containing the haplotype frequency spectrum of 30 dog brees for gene HSA2 located in Chromosome 13. Cell (i,j) of the matrix contains the number of occurences of the haplotype j in the breed i. A total of 124 haplotypes have been observed in the data so that the data.frame has 30 rows and 124 columns.
}
\value{
A dataset containing the haplotype frequency spectrum of 30 dog breeds for gene HSA2.}
\keyword{datasets}

